MyApps-13: Refactor remove JSON dependency from domain api.
