+++ app/src/main/java/xyz/torquato/myapps/api/web/model/BookError.kt
@@ -1,0 +1,5 @@
+package xyz.torquato.myapps.api.web.model
+
+data class BookError(
+    val message: String
+)
