+++ app/src/main/java/xyz/torquato/myapps/api/web/model/QueryResult.kt
@@ -1,13 +1,10 @@
 package xyz.torquato.myapps.api.web.model
 
-import org.json.JSONException
-import org.json.JSONObject
-
 sealed interface QueryResult {
 
     data object Empty: QueryResult
 
-    data class Valid(val data: JSONObject): QueryResult
+    data class Valid(val data: List<BookItem>): QueryResult
 
-    data class Error(val error: JSONException): QueryResult
+    data class Error(val error: BookError): QueryResult
 }
