+++ app/src/main/java/xyz/torquato/myapps/domain/web/content/model/QueryResultCollection.kt
@@ -8,29 +8,7 @@ sealed interface QueryResultCollection {
 
     data class Valid(
         val content: List<QueryResult.Valid>
-    ): QueryResultCollection {
-        data class BookItem(
-            var id: String,
-            var title: String,
-            var author: String,
-            var description: String,
-            var smallThumbnailUrl: String,
-            var largeThumbnailUrl: String,
-            var buyLink: String?
-        ) {
-            companion object {
-                fun empty() = BookItem(
-                    id = "",
-                    title = "",
-                    author = "",
-                    description = "",
-                    smallThumbnailUrl = "",
-                    largeThumbnailUrl = "",
-                    buyLink = null
-                )
-            }
-        }
-    }
+    ): QueryResultCollection
 
     data object Error: QueryResultCollection
